+++ src/models/rating/score_rating/types.rs
@@ -1,3 +1,4 @@
+use bigdecimal::BigDecimal;
 use chrono::NaiveDateTime;
 use validator::Validate;
 
@@ -15,12 +16,11 @@ pub struct ScoreRatingRow {
 
     /// Rating value for the score.
     /// Must be a positive decimal value.
-    #[validate(range(min = 0.01, message = "Rating must be positive"))]
-    pub rating: f64,
+    pub rating: BigDecimal,
 
     /// Type of rating system used.
     /// Must be one of: 'osu', 'etterna', 'quaver', 'malody', 'interlude'.
-    #[validate(custom = "validate_rating_type")]
+    #[validate(custom(function = "validate_rating_type"))]
     pub rating_type: String,
 
     /// Timestamp when the score rating was created.
@@ -33,4 +33,3 @@ fn validate_rating_type(rating_type: &str) -> Result<(), validator::ValidationEr
         _ => Err(validator::ValidationError::new("invalid_rating_type")),
     }
 }
-
